+++ linux/keyman-system-service/build.sh
@@ -3,7 +3,7 @@
 ## START STANDARD BUILD SCRIPT INCLUDE
 # adjust relative paths as necessary
 THIS_SCRIPT="$(readlink -f "${BASH_SOURCE[0]}")"
-. "${THIS_SCRIPT%/*}/../../resources/build/build-utils.sh"
+. "${THIS_SCRIPT%/*}/../../resources/build/builder.inc.sh"
 ## END STANDARD BUILD SCRIPT INCLUDE
 
 ################################ Main script ################################
@@ -22,8 +22,6 @@ builder_describe \
   "--coverage                capture test coverage"
 
 builder_parse "$@"
-
-cd "$THIS_SCRIPT_PATH/"
 
 if builder_is_debug_build; then
   MESON_TARGET=debug