+++ HISTORY.md
@@ -1,5 +1,14 @@
 # Keyman Version History
 
+## 18.0.29 alpha 2024-05-06
+
+* chore(web): Improve web/test.sh script (#11355)
+* chore(linux): Fix typo (#11356)
+
+## 18.0.28 alpha 2024-05-04
+
+* chore(common): maintenance on build scripts - cd (#11329)
+
 ## 18.0.27 alpha 2024-05-03
 
 * chore(linux): Adjust distros for GHA (#11333)